refactor: extract active-state todo editing helpers
